Make max buffer size a multiple of 3968 bytes
Rename 2 defines in the process.

The AN232B-03 spells clearly that the buffer should be a multiple of the
available user data bytes in the endpoint buffer.

By default, the FTDI endpoint buffer is 4096 bytes, and USB packets are
64 bytes, of which 2 bytes are control bytes not available for user data.
The resulting user-available buffer space is thus 3968 bytes per buffer.
Each buffer is sent as a single USB transaction. To minimize the number
of transactions, user data should be packed in multiple of 3968 bytes.

On my test machine, this change resulted in c.20% CPU usage reduction.