Move buffer flipping to output thread
This change improves output bandwidth and reduces lock contention.

It also breaks synchronization with read() calls to the devices, so the
exSID_clkdread() is now only valid when the driver is built unthreaded.
If exSID_clkdread() is called in threaded operation, it will only clock
the cycles but will not perform read from the device.

This change also fixes the internal read() routines so it works as
expected with libftdi, meaning that it now correctly blocks until all
the requested data has been fetched.

This means that the driver is now operating correctly with libftdi and
there is no reason to favor ftd2xx.